update: config
