+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2020-2021++  K8sCat <k8scat@gmail.com>
+Copyright (c) 2020 K8sCat and The Gigrator Authors
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
